Add UI elements and backing features for Logarithmic scaling
This commit adds the UI elements and the structure changes necessary to allow the user to choose
X axis scaling and transmit that option to the functional code. The options are "Linear" which is
the current capability, Log10, and Log2 (the former being very common, the latter being a bit more
natural for aural use, but not as common). The functional code is added in the next commit.