POC implementation of logarithmic X-axis scaling
Based upon user selection, the Spectroscope can now also display in either
Log2 scaling or Log10 scaling. A lot of the changes are simply related to trying
to understand the algorithm that created the linear scaling - i.e. giving variables
useful names and adding comments. The commit provides the basic display (which is
not filled, but "stroked"). Obviously this isn't very pretty because there are far fewer
samples in the 10-100 range than there are in, say, the 1,000 to
10,000 range, but a logarithmic scale devotes the same space to each - however, it
still does allow the common form to be recognized and for easier spotting of  filter
impact.

This is incomplete - zoom, cursor position, and events are all "wrong" at the moment,
and actual positioning of the plotted data also seems a bit off - but it gives a taste of
what is to come.