Fix spacing and cursor; refactor for a LogMode type
Introduce a LogMode type that centralizes the complexities and
commonalities of dealing for logarithmic scales. Make sure the
spacing makes maximum use of the space given, and that the
cursor reporting is accurate.

We do now have a slight problem! While the spectroscope
logarithmic display is deadly accurate, there is a "j" toward the
end of the function that we can no longer derive from the x axis...
we should be able to get an estimate, but I've not done that yet.